+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * Added `__serialize()` and `__unserialize()` methods to the `EntityId` interface.
 * Added native type hints to the `Statement` and `StatementList` classes
 * Added `strict_types=1` to `Statement.php`, `StatementList.php`, and related test files
+* Removed support for repository names in entity IDs (e.g. `foo:Q1234`)
 
 ## Version 9.6.1 (2021-04-01)